# frontend/src/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Dataset Request Desk API",
        "version": "1.0.0",
        "description": "Internal API for dataset requests, episodes and fulfilment."
    },
    "paths": {
...
        "/api/requests/{id}/assignments/": {
            "get": {
                "operationId": "requests_assignments_list",
                "description": "Episodes assigned to the request. Clients see what is currently assigned to their own.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "history",
                        "schema": {
                            "enum": [
                                "true",
                                "false"
                            ],
                            "type": "string",
                            "minLength": 1
                        },
                        "description": "Staff only: include released assignments.\n\n* `true` - true\n* `false` - false"
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "A UUID string identifying this dataset request.",
                        "required": true
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "A page number within the paginated result set.",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page_size",
                        "required": false,
                        "in": "query",
                        "description": "Number of results to return per page.",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                        }
                    }
                ],
                "tags": [
                    "requests"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedAssignmentHistoryList"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid input; `details` lists the problems per field.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not signed in, or the access token is invalid or expired.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Signed in, but your role may not do this.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found, or not visible to you (or a page past the end).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests; retry after the number of seconds in `Retry-After`.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "requests_assignments_create",
                "description": "Assign episodes (operators and admins). All or nothing: any problem assigns none.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Idempotency-Key",
                        "schema": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255,
                            "pattern": "^[ -~]+$"
                        },
                        "description": "Optional, any unique string (e.g. a UUID). A retry with the same key and body replays the first response instead of doing the work again. Kept for 24 hours."
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "A UUID string identifying this dataset request.",
                        "required": true
                    }
                ],
                "tags": [
                    "requests"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignEpisodesRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssignResult"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid input; `details` lists the problems per field.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not signed in, or the access token is invalid or expired.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Signed in, but your role may not do this.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found, or not visible to you (or a page past the end).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The request is not in progress, or an episode is assigned elsewhere. Or: a request with this Idempotency-Key is still being processed."
                    },
                    "422": {
                        "description": "This Idempotency-Key was already used for a different request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests; retry after the number of seconds in `Retry-After`.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/requests/{id}/events/": {
            "get": {
                "operationId": "requests_events_list",
                "description": "Status history, oldest first, paginated (it grows with every rework). Append-only.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "A UUID string identifying this dataset request.",
                        "required": true
                    },
                    {
                        "name": "page",
                        "required": false,
                        "in": "query",
                        "description": "A page number within the paginated result set.",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page_size",
                        "required": false,
                        "in": "query",
                        "description": "Number of results to return per page.",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                        }
                    }
                ],
                "tags": [
                    "requests"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedRequestEventList"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid input; `details` lists the problems per field.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not signed in, or the access token is invalid or expired.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Signed in, but your role may not do this.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found, or not visible to you (or a page past the end).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests; retry after the number of seconds in `Retry-After`.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/requests/{id}/transitions/": {
            "post": {
                "operationId": "requests_transitions_create",
                "description": "Move the request through its workflow. Allowed steps depend on status and role.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Idempotency-Key",
                        "schema": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255,
                            "pattern": "^[ -~]+$"
                        },
                        "description": "Optional, any unique string (e.g. a UUID). A retry with the same key and body replays the first response instead of doing the work again. Kept for 24 hours."
                    },
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "description": "A UUID string identifying this dataset request.",
                        "required": true
                    }
                ],
                "tags": [
                    "requests"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransitionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DatasetRequest"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid input; `details` lists the problems per field.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not signed in, or the access token is invalid or expired.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Signed in, but your role may not do this.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found, or not visible to you (or a page past the end).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "Not a valid step from the current status, or too few episodes assigned. Or: a request with this Idempotency-Key is still being processed."
                    },
                    "422": {
                        "description": "This Idempotency-Key was already used for a different request.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests; retry after the number of seconds in `Retry-After`.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AssignEpisodesRequest": {
                "type": "object",
                "properties": {
                    "episode_ids": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 32
                        },
                        "maxItems": 500
                    }
                },
                "required": [
                    "episode_ids"
                ]
            },
            "AssignResult": {
                "type": "object",
                "properties": {
                    "assigned": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "assigned_count": {
                        "type": "integer"
                    },
                    "episodes_requested": {
                        "type": "integer"
                    }
                },
                "required": [
                    "assigned",
                    "assigned_count",
                    "episodes_requested"
                ]
            },
...
            "DatasetRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "client": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RequestClient"
                            }
                        ],
                        "readOnly": true
                    },
                    "task_name": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "episodes_requested": {
                        "type": "integer",
                        "maximum": 1000000,
                        "minimum": 1
                    },
                    "assigned_count": {
                        "type": "integer",
                        "readOnly": true,
                        "description": "Episodes currently assigned to the request."
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 2000
                    },
                    "status": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RequestStatus"
                            }
                        ],
                        "readOnly": true
                    },
                    "status_changed_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "assigned_count",
                    "client",
                    "created_at",
                    "deadline",
                    "episodes_requested",
                    "id",
                    "status",
                    "status_changed_at",
                    "task_name"
                ]
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorBody"
                    }
                },
                "required": [
                    "error"
                ]
            },
...
            "PaginatedAssignmentHistoryList": {
                "type": "object",
                "required": [
                    "count",
                    "results"
                ],
                "properties": {
                    "count": {
                        "type": "integer",
                        "example": 123
                    },
                    "next": {
                        "type": "string",
                        "nullable": true,
                        "format": "uri",
                        "example": "http://api.example.org/accounts/?page=4"
                    },
                    "previous": {
                        "type": "string",
                        "nullable": true,
                        "format": "uri",
                        "example": "http://api.example.org/accounts/?page=2"
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AssignmentHistory"
                        }
                    }
                }
            },
...
            "PaginatedRequestEventList": {
                "type": "object",
                "required": [
                    "count",
                    "results"
                ],
                "properties": {
                    "count": {
                        "type": "integer",
                        "example": 123
                    },
                    "next": {
                        "type": "string",
                        "nullable": true,
                        "format": "uri",
                        "example": "http://api.example.org/accounts/?page=4"
                    },
                    "previous": {
                        "type": "string",
                        "nullable": true,
                        "format": "uri",
                        "example": "http://api.example.org/accounts/?page=2"
                    },
                    "results": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RequestEvent"
                        }
                    }
                }
            },
...
            "TransitionRequest": {
                "type": "object",
                "properties": {
                    "to_status": {
                        "$ref": "#/components/schemas/RequestStatus"
                    },
                    "comment": {
                        "type": "string",
                        "default": "",
                        "maxLength": 2000
                    }
                },
                "required": [
                    "to_status"
                ]
            },
...
        },
        "securitySchemes": {
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}